backend/requirements.txt
aioredis==1.3.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
click==8.1.3
colorama==0.4.6
colorspacious==1.1.2
constantly==15.1.0
cryptography==40.0.1
cycler==0.11.0
daphne==3.0.2
Django==3.0.10
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.11.1
fonttools==4.34.4
h2==4.1.0
hiredis==2.2.2
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
JsonForm==0.0.2
jsonschema==4.17.3
JsonSir==0.0.2
kiwisolver==1.4.4
matplotlib==3.5.2
msgpack==1.0.5
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
plottools==0.2.0
priority==1.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
pytz==2023.3
PyYAML==6.0
redis==4.5.4
Resource==0.2.1
rq==1.13.0
scipy==1.8.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
zope.interface==6.0